refactor: use transform API to support node addons
